fix: transaction.generate_transaction() gains access to wallet
